Moved Day Period to another file
